src/test/java/com/example/studyschedule/service/study/StudyServiceTest.java
package com.example.studyschedule.service.study;

import com.example.studyschedule.entity.member.Member;
import com.example.studyschedule.entity.study.Study;
import com.example.studyschedule.enums.IsUse;
import com.example.studyschedule.model.dto.study.StudyDto;
import com.example.studyschedule.repository.member.MemberRepository;
import com.example.studyschedule.repository.study.StudyRepository;
import org.junit.jupiter.api.AfterEach;
import org.junit.jupiter.api.DisplayName;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.transaction.annotation.Transactional;

import java.util.List;

import static org.assertj.core.api.Assertions.assertThat;
import static org.junit.jupiter.api.Assertions.*;

@SpringBootTest
@Transactional
class StudyServiceTest {

    @Autowired
    StudyService service;

    @Autowired
    StudyRepository studyRepository;

    @Autowired
    MemberRepository memberRepository;

    @Test
    @DisplayName("현재 사용 가능하고 공개된 스터디가 조회된다.")
    void getPublicStudyListTest() {
        Member savedMember = memberRepository.save(new Member("Test", 100));
        studyRepository.save(Study.ofPublicStudy(savedMember, "Study Test", 10L, IsUse.Y));

        List<StudyDto> response = service.getPublicStudyList();

        assertThat(response).hasSize(1);
        assertThat(response.get(0).getStudyName()).isEqualTo("Study Test");
        assertThat(response.get(0).getLeaderName()).isEqualTo("Test");
    }

}